packages/server/requirements.txt
# FastAPI and async server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.1.0
slowapi>=0.1.9

# MCP SDK for official protocol support
mcp>=1.0.0

# Schema validation (updated for Python 3.13 compatibility)
pydantic>=2.10.0
pydantic-settings>=2.4.0

# Async HTTP client (>=0.27 required by mcp)
httpx>=0.27.0

# Logging
loguru==0.7.2

# Caching (optional)
redis[hiredis]==5.0.1

# External API integrations
requests==2.31.0
todoist-api-python==2.1.4
python-dateutil==2.8.2

# Google Calendar API
google-api-python-client>=2.140.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
pytz==2025.2  # Timezone handling for smart scheduling

# Environment file loading (optional)
python-dotenv==1.0.0
